voteの実装をservicesに移動
